requirements.txt
Pillow
freetype-py
pyyaml
